Apply govuk-body styles
- Tedious to apply class `.govuk-body` to all <p> tags,
therefore we extend the selector govuk-body-m
and apply it to all <p> within the <main> tag.

- Some long form content isn't semantically correct,
it is tedious to tackle all of them at once in this task,
therefore we apply typography styles to all content within <main>

- We have different webpack entry files for the old and new stylesheets,
because of this, in conjunction with the typography changes,
we need to define the frontend-font-url location as to not
create any routing errors when requesting the font files.

This is temporary and can be removed once overall migration work is complete,
as we should then have a single entry file for our assets.